+++ classes/ticket_processor_class.rb
@@ -1,34 +1,5 @@
-require 'sendgrid-ruby'
-
-class EmailClient
-  include SendGrid
-
-  def send_email(from, to, subject, message)
-    from = Email.new(email: from)
-    to = Email.new(email: to)
-    content = Content.new(type: 'text/plain', value: message)
-    mail = Mail.new(from, subject, to, content)
-    sg = SendGrid::API.new(api_key: ENV.fetch("SENDGRID_API_KEY"))
-    response = sg.client.mail._('send').post(request_body: mail.to_json)
-    puts response.status_code
-    puts response.body
-    puts response.headers
-  end
-end
-
-class TicketPrinter
-  def print_ticket(name, cost, time)
-    puts name
-    puts "#{cost} zł"
-    puts time
-    puts seat_selection
-  end
-
-  private
-  def seat_selection
-    rand 1..150
-  end
-end
+require './email_client.rb'
+require './ticket_printer.rb'
 
 class CinemaTicket
   def initialize(movie_name, ticket_cost, screening_time)
@@ -58,7 +29,14 @@ class TheaterTicket
     message = 'Enjoy spectacle'
     EmailClient.new.send_email(from, to, subject, message)
   end
+
+  def print
+    TicketPrinter.new.print_ticket(@show_name, @ticket_cost, @start_time)
+  end
 end
 
 cinema_ticket = CinemaTicket.new('Titanic', 20, '18:00')
 cinema_ticket.send_by_email('test@test.com')
+
+theater_ticket = TheaterTicket.new('Titanic', 20, '18:00')
+theater_ticket.print
